Refactor event handlers into modular structure
