+++ src/features/posts/index.js
@@ -1,0 +1,11 @@
+import { initPostHandlers } from './actions.js';
+import { initFilterHandlers, applyFilterAndRender } from './filters.js';
+import { initPreviewHandlers } from './preview.js';
+
+export function initPosts() {
+  initPostHandlers();
+  initFilterHandlers();
+  initPreviewHandlers();
+}
+
+export { applyFilterAndRender };
