+++ src/main.js
@@ -11,11 +11,11 @@ import { GQL_QUERY, FETCH_CONTACTS_QUERY } from './api/queries.js';
 import { safeArray } from './utils/formatter.js';
 import { buildTree } from './ui/render.js';
 import { mergeLists } from './utils/merge.js';
-import { applyFilterAndRender } from './features/posts/events.js';
+import { initPosts, applyFilterAndRender } from './features/posts/index.js';
 import { fetchGraphQL } from './api/fetch.js';
 import { tribute } from './utils/tribute.js';
 import { initFilePond, resumeAudioContext } from './utils/filePond.js';
-import { initNotifications } from './features/notifications/events.js';
+import { initNotifications } from './features/notifications/index.js';
 import './features/uploads/handlers.js';
 
 function flattenComments(posts) {
@@ -138,6 +138,7 @@ window.addEventListener('DOMContentLoaded', () => {
       image: c.Profile_Image,
     }));
   });
+  initPosts();
   initFilePond();
   connect();
   initNotifications();
